Weather now renders on park selection on the dropdown. Weather is based on the zip of the park.
