+++ scripts/main.js
@@ -4,9 +4,8 @@ import { getAttractions } from "./attractions/AttractionProvider.js"
 import AttractionSelect from "./attractions/AttractionSelect.js"
 import { getEateries } from "./eateries/EateryProvider.js"
 import { eaterySelect } from "./eateries/EaterySelect.js"
-import { getWeather } from "./weather/WeatherProvider.js"
 import  "./eateries/EateryCard.js"
-import { FilterWeather } from "./weather/WeatherList.js"
+import "./weather/WeatherList.js"
 import "./attractions/AttractionCard.js"
 import { eateryDialogButton } from "./eateries/EateryDetailDialog.js"
 import "./parks/ParkCard.js"
@@ -24,7 +23,5 @@ getEateries()
     .then(eaterySelect)
     .then(eateryDialogButton)
 
-getWeather()
-    .then(FilterWeather)
