+++ scripts/weather/WeatherList.js
@@ -1,7 +1,9 @@
-import { useWeather } from "./WeatherProvider.js"
+import { useWeather, getWeather } from "./WeatherProvider.js"
 import { Weather } from "./Weather.js"
+import { useParks } from "../parks/ParkProvider.js"
 
 const contentTarget = document.querySelector(".weather")
+const eventHub = document.querySelector(".container")
 
 export const FilterWeather = () => {
 
@@ -18,9 +20,7 @@ export const FilterWeather = () => {
     render(weatherToDisplay)
 }
 
-
 // Render the weather objects
-
 export const render = weatherToRender => {
     contentTarget.innerHTML = weatherToRender.map(
         (weatherObject) => {
@@ -28,3 +28,23 @@ export const render = weatherToRender => {
         }
         ).join("")
     }
+    // Listen for changes to the weather state
+    eventHub.addEventListener("weatherStateChanged", customEvent => {
+    FilterWeather()
+})
+
+// Listen for the park state to be changed, shout the zip of the chosen park to other modules
+eventHub.addEventListener("parkChosen", event => {
+    const allParks = useParks()
+    const thePark = event.detail.park
+    const chosenPark = allParks.find(
+        (currentPark) => {
+            if (currentPark.name === thePark) {
+                return currentPark
+            }
+        }
+    )
+    let currentZip = chosenPark.addresses[0].postalCode
+    getWeather(currentZip)
+}
+)
